fix: align Windows recovery validation
